requirements/base.txt
accelerate==0.20.3
CairoSVG==2.7.0
datasets==2.12.0
diffusers==0.16.1
Pillow==9.5.0
qrcode==7.4.2 
torch==2.0.1
torchvision==0.15.2
transformers==4.30.1
python-dotenv==1.0.0 